+++ spinetoolbox/widgets/settings_widget.py
@@ -460,6 +460,7 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
         datetime = int(self._qsettings.value("appSettings/dateTime", defaultValue="2"))
         delete_data = int(self._qsettings.value("appSettings/deleteData", defaultValue="0"))
         use_experimental_engine = self._qsettings.value("appSettings/useExperimentalEngine", defaultValue="true")
+        custom_open_project_dialog = self._qsettings.value("appSettings/customOpenProjectDialog", defaultValue="true")
         smooth_zoom = self._qsettings.value("appSettings/smoothZoom", defaultValue="false")
         curved_links = self._qsettings.value("appSettings/curvedLinks", defaultValue="false")
         data_flow_anim_dur = int(self._qsettings.value("appSettings/dataFlowAnimationDuration", defaultValue="100"))
@@ -490,6 +491,8 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
             self.ui.checkBox_delete_data.setCheckState(Qt.Checked)
         if use_experimental_engine == "true":
             self.ui.checkBox_use_experimental_engine.setCheckState(Qt.Checked)
+        if custom_open_project_dialog == "true":
+            self.ui.checkBox_custom_open_project_dialog.setCheckState(Qt.Checked)
         if smooth_zoom == "true":
             self.ui.checkBox_use_smooth_zoom.setCheckState(Qt.Checked)
         if curved_links == "true":
@@ -568,6 +571,8 @@ class SettingsWidget(SpineDBEditorSettingsMixin, SettingsWidgetBase):
         self._qsettings.setValue("appSettings/deleteData", delete_data)
         use_experimental_engine = "true" if int(self.ui.checkBox_use_experimental_engine.checkState()) else "false"
         self._qsettings.setValue("appSettings/useExperimentalEngine", use_experimental_engine)
+        custom_open_project_dial = "true" if int(self.ui.checkBox_custom_open_project_dialog.checkState()) else "false"
+        self._qsettings.setValue("appSettings/customOpenProjectDialog", custom_open_project_dial)
         smooth_zoom = "true" if int(self.ui.checkBox_use_smooth_zoom.checkState()) else "false"
         self._qsettings.setValue("appSettings/smoothZoom", smooth_zoom)
         curved_links = "true" if int(self.ui.checkBox_use_curved_links.checkState()) else "false"